Separate test files for each sanitizer
I noticed a typo (Sanizer vs Sanitizer) in the core test file and
decided that, with this fix, it would also be better to have separate
test files so that it is easier to create new sanitizers.
